+++ controllers/machines.js
@@ -3,8 +3,7 @@ const Organization = require('../models/organization');
 const util = require('../services/util');
 
 exports.getAll = function(req, res) {
-	let query = util.filter.query(req.query, ['organization']);
-	Machine.find(query, (err, machines) => {
+	Machine.find({'organization' : req.user.organization}, (err, machines) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
 		res.send({ machines });
 	});
@@ -13,7 +12,9 @@ exports.getAll = function(req, res) {
 exports.get = function(req, res) {
 	Machine.findById(req.params.id, (err, machine) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
-		if (!machine) return res.status(404).send(util.norm.errors({ message: 'Machine not found' }));
+		if (!machine || String(machine.organization) != req.user.organization) {
+			return res.status(404).send(util.norm.errors({ message: 'Machine not found' }));
+		}
 		res.send({ machine });
 	});
 }
@@ -21,16 +22,19 @@ exports.get = function(req, res) {
 exports.post = function(req, res) {
 	var machine = new Machine(req.body.machine);
 
-	// TODO: Verify data before inserting
-
 	// Assign current user's organization if null
 	machine.organization = machine.organization || req.user.organization;
+
+	// TODO: Verify data before inserting
+	if (String(machine.organization) != req.user.organization) {
+		return res.status(404).send(util.norm.errors({ message: 'Cannot create machine outside of organization' }));
+	}
 
 	// Try to find the organization
 	Organization.findById(machine.organization, (err, organization) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
 		if (!organization) return res.status(404).send(util.norm.errors({ message: 'Organization not found' }));
-		//Save machine
+		// Save machine
 		machine.save(err => {
 			if (err) return res.status(500).send(util.norm.errors(err));
 			res.send({ machine });
@@ -42,12 +46,15 @@ exports.update = function(req, res) {
 	let updatedMachine = req.body.machine;
 
 	// TODO: Verify changes before updating
+	if (updatedMachine.organization != req.user.organization) {
+		return res.status(404).send(util.norm.errors({ message: 'Cannot update machine outside of organization' }));
+	}
 
-	//Try to find the organization
-	Organization.findById(machine.organization, (err, organization) => {
+	// Try to find the organization
+	Organization.findById(updatedMachine.organization, (err, organization) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
 		if (!organization) return res.status(404).send(util.norm.errors({ message: 'Organization not found' }));
-		//Save machine
+		// Save machine
 		Machine.findByIdAndUpdate(req.params.id, updatedMachine, { new: true }, (err, machine) => {
 			if (err) return res.status(500).send(util.norm.errors(err));
 			res.send({ machine });
@@ -56,9 +63,16 @@ exports.update = function(req, res) {
 }
 
 exports.delete = function(req, res) {
-	Machine.findByIdAndRemove(req.params.id, (err, deleted) => {
+	User.findById(req.params.id, (err, user) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
-		if (!deleted) return res.status(404).send(util.norm.errors({ message: 'Record not found' }))
-		res.sendStatus(200);
-	});
+		if (!user || String(user.organization) != req.user.organization) {
+			return res.status(404).send(util.norm.errors({ message: 'Cannot delete machine outside of organization' }));
+		}
+
+		Machine.findByIdAndRemove(req.params.id, (err, deleted) => {
+			if (err) return res.status(500).send(util.norm.errors(err));
+			if (!deleted) return res.status(404).send(util.norm.errors({ message: 'Record not found' }))
+			res.sendStatus(200);
+		});
+	});	
 }
