+++ controllers/users.js
@@ -8,9 +8,8 @@ exports.me = function(req, res) {
 }
 
 exports.getAll = function(req, res) {
-	// Allow the client to query users' by organization
-	let query = util.filter.query(req.query, ['organization']);
-	User.find(query, (err, users) => {
+	// Allow the client to query only the user's organization
+	User.find({'organization' : req.user.organization}, (err, users) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
 		res.send({ users });
 	});
@@ -19,7 +18,9 @@ exports.getAll = function(req, res) {
 exports.get = function(req, res) {
 	User.findById(req.params.id, (err, user) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
-		if (!user) return res.status(404).send(util.norm.errors({ message: 'User not found' }));
+		if (!user || String(user.organization) != req.user.organization) {
+			return res.status(404).send(util.norm.errors({ message: 'User not found' }));
+		}
 		res.send({ user });
 	});
 }
@@ -27,11 +28,14 @@ exports.get = function(req, res) {
 exports.post = function(req, res) {
 	let user = new User(req.body.user);
 
-	// TODO: Verify data before inserting
-
 	// Assign current user's organization if null
 	user.organization = user.organization || req.user.organization;
-
+	
+	// TODO: Verify data before inserting
+	if (String(user.organization) != req.user.organization) {
+		return res.status(404).send(util.norm.errors({ message: 'Cannot create user outside of organization' }));
+	}
+	
 	user.save(err => {
 		if (err) return res.status(500).send(util.norm.errors(err));
 		res.send({ user });
@@ -42,10 +46,15 @@ exports.update = function(req, res) {
 	let updatedUser = req.body.user;
 
 	// TODO: Verify changes before updating
+	if (!req.user.isAdmin && updatedUser._id != req.user._id) {
+		return res.status(404).send(util.norm.errors({ message: 'Cannot update another user' }));
+	}
 
 	User.findById(req.params.id, (err, user) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
-		if (!user) return res.status(404).send(util.norm.errors({ message: 'User not found' }));
+		if (!user || String(user.organization) != req.user.organization) {
+			return res.status(404).send(util.norm.errors({ message: 'User not found' }));
+		}
 
 		// Determine how to handle the message
 		let origin = util.checkOrigin(req.user, user._id);
@@ -58,9 +67,11 @@ exports.update = function(req, res) {
 			let oldUser = JSON.parse(JSON.stringify(user));
 			if (updatedUser.name) user.name = updatedUser.name;
 			if (updatedUser.email) user.email = updatedUser.email;
-			if (updatedUser.isAdmin) user.isAdmin = updatedUser.isAdmin;
+			if (req.user.isAdmin && updatedUser.isAdmin) user.isAdmin = updatedUser.isAdmin;
 			if (updatedUser.phoneNumber) user.phoneNumber = updatedUser.phoneNumber;
 			if (updatedUser.organization) user.organization = updatedUser.organization;
+			if (updatedUser.password) user.password = updatedUser.password;
+			if (updatedUser.googleAuthKey) user.googleAuthKey = updatedUser.googleAuthKey;
 
 			user.save((err, saved) => {
 				if (err) return res.status(500).send(util.norm.errors(err));
@@ -79,9 +90,16 @@ exports.update = function(req, res) {
 }
 
 exports.delete = function(req, res) {
-	User.findByIdAndRemove(req.params.id, (err, deleted) => {
+	User.findById(req.params.id, (err, user) => {
 		if (err) return res.status(500).send(util.norm.errors(err));
-		if (!deleted) return res.status(404).send(util.norm.errors({ message: 'Record not found' }))
-		res.sendStatus(200);
+		if (!user || String(user.organization) != req.user.organization) {
+			return res.status(404).send(util.norm.errors({ message: 'Cannot delete user outside of organization' }));
+		}
+
+		User.findByIdAndRemove(req.params.id, (err, deleted) => {
+			if (err) return res.status(500).send(util.norm.errors(err));
+			if (!deleted) return res.status(404).send(util.norm.errors({ message: 'Record not found' }))
+			res.sendStatus(200);
+		});
 	});
 }
